make interface of signature clean
